refactor: DataAPI delete ifs & split interfaces to modules
