+++ src/scripts/common/api/models/User.model.ts
@@ -1,0 +1,7 @@
+interface User {
+  name: string;
+  email: string;
+  password: string;
+}
+
+export default User;
