+++ src/scripts/common/api/models/UserSetting.model.ts
@@ -1,0 +1,6 @@
+interface UserSetting {
+  wordsPerDay: number;
+  optional: Record<string, unknown>;
+}
+
+export default UserSetting;
