+++ src/scripts/common/api/models/UserStatistic.model.ts
@@ -1,0 +1,6 @@
+interface UserStatistic {
+  learnedWords: number;
+  optional: Record<string, unknown>;
+}
+
+export default UserStatistic;
