+++ src/scripts/common/api/models/UserWord.model.ts
@@ -1,0 +1,6 @@
+interface UserWord {
+  difficulty: string;
+  optional: Record<string, unknown>;
+}
+
+export default UserWord;
